+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   - Fix opendkim config
   - Upgrade baseimage: light-baseimage:0.2.5
   - Add tls certificates for replication only
+  - Add opendmarc
 
 ## 0.1.7
   - Upgrade baseimage: light-baseimage:0.2.2